Added ServiceCategory APIs and Service Entity

+++ CleanZcare-Backend-2/src/main/java/com/cleanzcare/repository/ServiceCategoryRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cleanzcare.entities.ServiceCategory;
 
-public interface ServiceCategoryRepository extends JpaRepository<ServiceCategory, Integer> {
+public interface ServiceCategoryRepository extends JpaRepository<ServiceCategory, Long> {
 	boolean existsByCategoryName(String categoryName);
 
 }
